Increase retry tokens to support streams with lots of shards
We are seeing the quota error during `SubscribeToShard` because
Kinesis client is shared across go routines concurrently subscribing to
multiple shards. They are all sharing the retry token bucket which
has 500 tokens by default. Each retry of `SubscribeToShard` costs
5 tokens. In streams with large number of shards it's possible
that more than 100 `SubscribeToShard` calls are failing and therefore
runs out of retry tokens.

This change increases the token count to 1000000.

Once we verify the stability of this fix, we may introduce a mechanism
to dynamically workout the token count based on the number of shards
in the target.

Also includes:
- Go version is updated 1.19
- All dependencies are upgraded with go get -u

Fixes #7